requirements-social.txt
Flask==2.3.2
google-cloud-storage==2.10.0
google-generativeai==0.3.2
requests==2.31.0
gunicorn==21.2.0
